added nonlinear pseudo plastic hardening (not working yet)
